Moved all configs into config.go
